ingest/requirements.txt
requests==2.27.1
google.cloud.storage==2.2.1
google.cloud.logging==3.0.0